+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Version 0.3.1 - TBD
+### Added
+- Support for enums when they are defined separately (not inline in the property type of an entity)
 
 ## Version 0.3.0 - 2023-06-26
 ### Added